refactors complete, formatting issues resolved
