+++ lib/invoice.rb
@@ -2,7 +2,8 @@ require 'pry'
 require 'time'
 
 class Invoice
-attr_accessor :id, :customer_id, :merchant_id, :status, :created_at, :updated_at, :merchant, :customer, :items, :transactions, :total, :total_revenue, :invoice_items
+  attr_accessor :id, :customer_id, :merchant_id, :status, :created_at, :updated_at, :merchant, :customer, :items, :transactions, :total, :total_revenue, :invoice_items
+
   def initialize(attributes)
     @id = attributes[:id].to_i
     @customer_id = attributes[:customer_id].to_i
@@ -13,7 +14,6 @@ attr_accessor :id, :customer_id, :merchant_id, :status, :created_at, :updated_at
     @items = []
     @transactions = []
     @invoice_items = []
-    # @paid = true
   end
 
   def is_paid_in_full?
