+++ lib/merchant.rb
@@ -17,7 +17,7 @@ class Merchant
   def revenue
     invoice_revenue = []
     all_revenues = invoices.map do |invoice|
-      if invoice.is_paid_in_full? 
+      if invoice.is_paid_in_full?
         invoice.invoice_items.each do |item|
           invoice_revenue << item.unit_price * item.quantity
         end
