+++ lib/merchant_repository.rb
@@ -19,24 +19,6 @@ class MerchantRepository
     @all_merchants = []
   end
 
-  def from_csv(path)
-    if path.include? '.csv'
-    CSV.foreach(path, { headers: true, header_converters: :symbol}) do |data_row|
-      merchant = Merchant.new(data_row)
-      @all_merchants << merchant
-    end
-    else
-      populate_merchant_repo_with_hash(path)
-    end
-  end
-
-  def from_array(array)
-    array.each do |attributes|
-      merchant = Merchant.new(attributes)
-      @all_merchants << merchant
-    end
-  end
-
   def all
     @all_merchants
   end
@@ -56,5 +38,24 @@ class MerchantRepository
   def find_all_by_name(query_name)
     find_all_by_string(@all_merchants, query_name, :name)
   end
+
+  def from_csv(path)
+    if path.include? '.csv'
+      CSV.foreach(path, { headers: true, header_converters: :symbol}) do |data_row|
+        merchant = Merchant.new(data_row)
+        @all_merchants << merchant
+      end
+    else
+      populate_merchant_repo_with_hash(path)
+    end
+  end
+
+  def from_array(array)
+    array.each do |attributes|
+      merchant = Merchant.new(attributes)
+      @all_merchants << merchant
+    end
+  end
+
 
 end
