+++ lib/transaction_repository.rb
@@ -17,20 +17,6 @@ class TransactionRepository
     @all_transactions = []
   end
 
-  def from_csv(path)
-    CSV.foreach(path, { headers: true, header_converters: :symbol}) do |data_row|
-      transaction = Transaction.new(data_row)
-      @all_transactions << transaction
-    end
-  end
-
-  def from_array(array)
-    array.each do |attributes|
-      transaction = Transaction.new(attributes)
-      @all_transactions << transaction
-    end
-  end
-
   def all
     @all_transactions
   end
@@ -53,7 +39,7 @@ class TransactionRepository
 
   def find_all_by_result(result)
     @all_transactions.select do |element|
-       element.result.downcase == result ? element : nil
+      element.result.downcase == result ? element : nil
     end
   end
 
@@ -63,19 +49,32 @@ class TransactionRepository
     end
   end
 
+  def from_csv(path)
+    CSV.foreach(path, { headers: true, header_converters: :symbol}) do |data_row|
+      transaction = Transaction.new(data_row)
+      @all_transactions << transaction
+    end
+  end
+
+  def from_array(array)
+    array.each do |attributes|
+      transaction = Transaction.new(attributes)
+      @all_transactions << transaction
+    end
+  end
 
 end
 
 if __FILE__ == $0
-transaction = TransactionRepository.new
-transaction.from_csv("./data/transactions.csv")
+  transaction = TransactionRepository.new
+  transaction.from_csv("./data/transactions.csv")
 
   binding.pry
-engine.transactions.find_all_by_credit_card_number(4848466917766329)
-expect(expected.length).to eq 1
-expect(expected.first.class).to eq Transaction
-expect(expected.first.credit_card_number).to eq credit_card_number
+  engine.transactions.find_all_by_credit_card_number(4848466917766329)
+  expect(expected.length).to eq 1
+  expect(expected.first.class).to eq Transaction
+  expect(expected.first.credit_card_number).to eq credit_card_number
 
-credit_card_number = 4848466917766328
-expected = engine.transactions.find_all_by_credit_card_number(credit_card_number)
+  credit_card_number = 4848466917766328
+  expected = engine.transactions.find_all_by_credit_card_number(credit_card_number)
 end
